# src/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for authentication",
        "title": "Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/user/auth/health": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Проверка работоспособности сервиса",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/auth/signup/email": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.QEmailSignup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.ZEmailSignup"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.QEmailSignup": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.ZEmailSignup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-02-13 05:37:40.483836"
                },
                "id": {
                    "type": "string",
                    "example": "592af5b5-4f60-4ddd-b080-be674c86eda8"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-13 05:37:40.483836"
                }
            }
        }
    }
}